Cleaning - remove old bot
